bin/swx: Add option to rotate a container using cage
The problem:
I want to read PDFs and webpages on my laptop holding it upright.
Using `swaymsg output $laptop transform 90 clockwise` works BUT the
trackpad will now be unusable since it maps movements according to the
new screen orientation.

The solution:
Spawn the window inside cage [1] which allows us to rotate just that
container which will keep the input usable.

[1]: https://github.com/Hjdskes/cage